+++ src/app/publisher/dashboard/dashboard.component.ts
@@ -70,7 +70,7 @@ export class DashboardComponent extends HandleSubscription implements OnInit {
       )
       .subscribe(data => {
         this.barChartData[0].data = data.values;
-        this.barChartData[0].currentSeries = this.currentChartFilterSettings.currentSeries.label;
+        this.barChartData[0].currentSeries = chartFilterSettings.currentSeries.label;
         this.barChartLabels = data.timestamps.map(item => moment(item).format());
         this.barChartValue = data.total;
         this.barChartDifference = data.difference;