+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [1.4.8] - 2020-01-03
 ### Changed
 - Default direct link value
+- Limit chart requests
 ### Fixed
 - Long direct links