+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [1.5.0] - 2020-01-15
 ### Added
 - NowPayments integration
 - Additional button which adds advertisement
@@ -188,7 +190,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Publisher features (Sites & AdUnits)
 
 
-[Unreleased]: https://github.com/adshares/adpanel/compare/v1.4.6...develop
+[Unreleased]: https://github.com/adshares/adpanel/compare/v1.5.0...develop
+[1.5.0]: https://github.com/adshares/adpanel/compare/v1.4.8...v1.5.0
+[1.4.8]: https://github.com/adshares/adpanel/compare/v1.4.7...v1.4.8
+[1.4.7]: https://github.com/adshares/adpanel/compare/v1.4.6...v1.4.7
 [1.4.6]: https://github.com/adshares/adpanel/compare/v1.4.5...v1.4.6
 [1.4.5]: https://github.com/adshares/adpanel/compare/v1.4.4...v1.4.5
 [1.4.4]: https://github.com/adshares/adpanel/compare/v1.4.3...v1.4.4